Moved PreviewServer to use Kestrel.
